backend/pyproject.toml
[project]
name = "behemoth-backend"
version = "0.1.0"
description = "behemoth backend"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.8",
    "django-shortuuidfield>=0.1.3",
    "djangorestframework>=3.16.1",
    "djangorestframework-stubs[compatible-mypy]>=3.16.5",
    "python-dotenv>=1.2.1",
]

[dependency-groups]
dev = [
    "mypy>=1.18.1",
    "ruff>=0.14.7",
    "django-stubs>=5.2.7",
]

[tool.mypy]
strict = true
plugins = [
    "mypy_django_plugin.main",
    "mypy_drf_plugin.main",
]

[tool.django-stubs]
django_settings_module = "behemoth.settings"

[tool.ruff]
lint.select = ["ANN"]
